p/o/psreadinesscontroller: minor refactoring
